Updated Word_Game_Board
I deleted the old versions and uploaded this to be commited to phase 2 after review. I would like someone to review it. It's very explicit and could use some better logic.
As it stands:
Line 1
Line 2
Line 3...etc...
You'll need to take a look at the code to see why.